requirements.txt
construct==2.10.61
pillow==8.1.1
pyinstaller==4.2